taint/test/mochitest/test_fetch_response.html

<!DOCTYPE HTML>
<html>
  <head>
    <meta charset="utf-8">
    <title>Test HTML Taint Sinks</title>
    <script src="/tests/SimpleTest/SimpleTest.js"></script>
    <link rel="stylesheet" href="/tests/SimpleTest/test.css"/>
    <script>
      function check_tainted(str) {
        SimpleTest.ok(str.taint.length > 0, "Check tainted for: " + str);
      }
      let sink_names = [
        "innerHTML",
        "innerHTML"
      ];


      let i = 0;

      SimpleTest.waitForExplicitFinish();
      addEventListener("__taintreport", (report) => {

        let flow = report.detail.str.taint[0].flow;
        is(flow[1].operation, sink_names[i]);

        i += 1;
        if (i >= sink_names.length) {
          SimpleTest.finish();
        }
      }, false);

      async function startTest() {
        try {
          let response = await fetch(`http://mochi.test:8888/tests/taint/test/mochitest/fetch_server.sjs?text`)

          if (response.ok) {
            let text = await response.text();
            SimpleTest.info(text);
            check_tainted(text);
            document.body.innerHTML = text;
          } else {
            SimpleTest.is(true, false, 'Request failed with status: ' + response.status);
          }
        } catch(error
        ) {
          SimpleTest.info(error.message);
          SimpleTest.is(true, false, "Error");
        };
        try {
          let response = await fetch(`http://mochi.test:8888/tests/taint/test/mochitest/fetch_server.sjs?json`)

          if (response.ok) {
            let obj = await response.json();
            SimpleTest.info(JSON.stringify(obj));
            check_tainted(obj.txt);
            document.body.innerHTML = obj.txt;
          } else {
            SimpleTest.is(true, false, 'Request failed with status: ' + response.status);
          }
          return null;

        } catch(error
        ) {
          SimpleTest.info(error.message);
          SimpleTest.is(true, false, "Error");
        };
      }


    </script>
  </head>
  <body onload="startTest()">
    <p id="display"></p>
    <div id="content" style="display: none"></div>
    <p id="test"></p>
  </body>
</html>
